Move OfferViewUtil
