+++ desktop/src/main/java/bisq/desktop/main/overlays/windows/OfferDetailsWindow.java
@@ -22,7 +22,7 @@ import bisq.desktop.components.AutoTooltipButton;
 import bisq.desktop.components.BusyAnimation;
 import bisq.desktop.components.TitledGroupBg;
 import bisq.desktop.components.TxIdTextField;
-import bisq.desktop.main.offer.bisq_v1.OfferViewUtil;
+import bisq.desktop.main.offer.OfferViewUtil;
 import bisq.desktop.main.overlays.Overlay;
 import bisq.desktop.util.DisplayUtils;
 import bisq.desktop.util.GUIUtil;
